requirements.txt
langchain==1.0.3
langchain_google_genai==3.0.0
langgraph==1.0.2
python-dotenv==1.2.1
